goog/pip-requirements.txt
Django==1.8.1
djangorestframework==3.6.3
django-filter==1.0.4
requests==2.17.3
